# Speedup the render by removing function calls and reducing allocations