Results on actual sources:

Before:

```
Benchmark_testData/testdata/import.star-4               	  323349	      3826 ns/op	    2680 B/op	      47 allocs/op
Benchmark_testData/testdata/test_input_1.star-4         	   19461	     60488 ns/op	   36790 B/op	     641 allocs/op
Benchmark_testData/testdata/test_input_2.star-4         	   39580	     29471 ns/op	   18531 B/op	     344 allocs/op
```

After:

```
Benchmark_testData/testdata/import.star,_real_StringBuiler:_true-4          	 1000000	      1069 ns/op	     659 B/op	       4 allocs/op
Benchmark_testData/testdata/test_input_1.star,_real_StringBuiler:_true-4    	   64516	     18145 ns/op	   10598 B/op	      50 allocs/op
Benchmark_testData/testdata/test_input_2.star,_real_StringBuiler:_true-4    	  143785	      8459 ns/op	    4439 B/op	      20 allocs/op
```

Synthetic tests results:

Before:

```
Benchmark_huge_encapsulation_expr/1-4 	  140895	      8157 ns/op	    5504 B/op	      91 allocs/op
Benchmark_huge_encapsulation_expr/10-4         	   14613	     82219 ns/op	   55048 B/op	     910 allocs/op
Benchmark_huge_encapsulation_expr/100-4        	    1348	    886079 ns/op	  550500 B/op	    9100 allocs/op
Benchmark_huge_encapsulation_expr/1000-4       	      67	  15125048 ns/op	 5505504 B/op	   91008 allocs/op
Benchmark_huge_encapsulation_expr/10000-4      	       6	 179920299 ns/op	55042496 B/op	  910015 allocs/op
Benchmark_huge_encapsulation_expr/100000-4     	       1	3320775580 ns/op	550403752 B/op	 9100024 allocs/op
```

After:

```
Benchmark_huge_encapsulation_expr/1-4 	  743452	      1599 ns/op	      80 B/op	       2 allocs/op
Benchmark_huge_encapsulation_expr/10-4         	   80792	     15120 ns/op	      80 B/op	       2 allocs/op
Benchmark_huge_encapsulation_expr/100-4        	    7987	    150696 ns/op	      80 B/op	       2 allocs/op
Benchmark_huge_encapsulation_expr/1000-4       	     793	   1515312 ns/op	      80 B/op	       2 allocs/op
Benchmark_huge_encapsulation_expr/10000-4      	      66	  15756972 ns/op	      80 B/op	       2 allocs/op
Benchmark_huge_encapsulation_expr/100000-4     	       6	 169947340 ns/op	      80 B/op	       2 allocs/op
```


Changes in statement rendering, synthetic tests on huge indents may perform worse due to amount of whitespace writes.

Before:
```
Benchmark_huge_encapsulation_stmt/1-4 	  166652	      6750 ns/op	    4432 B/op	      74 allocs/op
Benchmark_huge_encapsulation_stmt/10-4         	   16690	     71170 ns/op	   49575 B/op	     740 allocs/op
Benchmark_huge_encapsulation_stmt/100-4        	    1356	    928818 ns/op	  983673 B/op	    7400 allocs/op
Benchmark_huge_encapsulation_stmt/1000-4       	      40	  27667112 ns/op	59908637 B/op	   74018 allocs/op
Benchmark_huge_encapsulation_stmt/10000-4      	       1	4241942485 ns/op	5614343008 B/op	  740792 allocs/op
```

After
```
Benchmark_huge_encapsulation_stmt/1-4 	  701206	      1729 ns/op	      96 B/op	       2 allocs/op
Benchmark_huge_encapsulation_stmt/10-4         	   39711	     27884 ns/op	      96 B/op	       2 allocs/op
Benchmark_huge_encapsulation_stmt/100-4        	     855	   1412810 ns/op	      96 B/op	       2 allocs/op
Benchmark_huge_encapsulation_stmt/1000-4       	       8	 125629725 ns/op	      96 B/op	       2 allocs/op
Benchmark_huge_encapsulation_stmt/10000-4      	       1	12446535444 ns/op	      96 B/op	       2 allocs/op
```